// src/types/manufacturer.ts
export interface Manufacturer {
  contact: string;      // Company/Individual name
  address1: string;     // Address line 1
  address2: string;     // Address line 2
  city: string;
  postalCode: string;
  provinceState: string;
  country: string;
  phone: string;
  email: string;
  isDefault: boolean;
};